Move int, float and Decimal tests to tests/types/
Also split the corresponding cases out of the `test_default_validators`
parameterization and replace `conint()`/`confloat()`/`condecimal()`
usages with the explicit annotated forms.
