+++ tests/test_types.py
@@ -396,81 +396,6 @@ def test_import_string_thing_with_name() -> None:
     assert import_things.model_dump_json() == '{"obj":{"name":"foo"}}'
 
 
-def test_decimal():
-    class Model(BaseModel):
-        v: Decimal
-
-    m = Model(v='1.234')
-    assert m.v == Decimal('1.234')
-    assert isinstance(m.v, Decimal)
-    assert m.model_dump() == {'v': Decimal('1.234')}
-
-
-def test_decimal_constraint_coerced() -> None:
-    ta = TypeAdapter(Annotated[Decimal, Field(gt=2)])
-
-    with pytest.raises(ValidationError):
-        ta.validate_python(Decimal(0))
-
-
-def test_decimal_allow_inf():
-    class MyModel(BaseModel):
-        value: Annotated[Decimal, AllowInfNan(True)]
-
-    m = MyModel(value='inf')
-    assert m.value == Decimal('inf')
-
-    m = MyModel(value=Decimal('inf'))
-    assert m.value == Decimal('inf')
-
-
-def test_decimal_dont_allow_inf():
-    class MyModel(BaseModel):
-        value: Decimal
-
-    with pytest.raises(ValidationError, match=r'Input should be a finite number \[type=finite_number'):
-        MyModel(value='inf')
-    with pytest.raises(ValidationError, match=r'Input should be a finite number \[type=finite_number'):
-        MyModel(value=Decimal('inf'))
-
-
-def test_decimal_strict():
-    class Model(BaseModel):
-        v: Decimal
-
-        model_config = ConfigDict(strict=True)
-
-    with pytest.raises(ValidationError) as exc_info:
-        Model(v=1.23)
-
-    # insert_assert(exc_info.value.errors(include_url=False))
-    assert exc_info.value.errors(include_url=False) == [
-        {
-            'type': 'is_instance_of',
-            'loc': ('v',),
-            'msg': 'Input should be an instance of Decimal',
-            'input': 1.23,
-            'ctx': {'class': 'Decimal'},
-        }
-    ]
-
-    v = Decimal(1.23)
-    assert Model(v=v).v == v
-    assert Model(v=v).model_dump() == {'v': v}
-
-    assert Model.model_validate_json('{"v": "1.23"}').v == Decimal('1.23')
-
-
-def test_decimal_precision() -> None:
-    ta = TypeAdapter(Decimal)
-
-    num = f'{1234567890 * 100}.{1234567890 * 100}'
-
-    expected = Decimal(num)
-    assert ta.validate_python(num) == expected
-    assert ta.validate_json(f'"{num}"') == expected
-
-
 def test_strict_date():
     class Model(BaseModel):
         v: Annotated[date, Field(strict=True)]
@@ -569,10 +494,7 @@ def test_strict_timedelta():
 def check_model_fixture():
     class CheckModel(BaseModel):
         bool_check: bool = True
-        int_check: int = 1
-        float_check: float = 1.0
         uuid_check: UUID = UUID('7bd00d58-6485-4ca6-b889-3da6d8df3ee4')
-        decimal_check: condecimal(allow_inf_nan=False) = Decimal('42.24')
         date_check: date = date(2017, 5, 5)
         datetime_check: datetime = datetime(2017, 5, 5, 10, 10, 10)
         time_check: time = time(10, 10, 10)
@@ -640,41 +562,12 @@ class BoolCastable:
         ('bool_check', Decimal(2), ValidationError),
         ('bool_check', b'\x81', ValidationError),
         ('bool_check', BoolCastable(), ValidationError),
-        pytest.param('int_check', 1, 1, id='int_check-1_int-1_int'),
-        ('int_check', 1.0, 1),
-        ('int_check', 1.9, ValidationError),
-        ('int_check', Decimal(1), 1),
-        ('int_check', Decimal(1.9), ValidationError),
-        pytest.param('int_check', '1', 1, id='int_check-1_str-1_int'),
-        pytest.param('int_check', '1.9', ValidationError, id='int_check-1.9_str-ValidationError'),
-        ('int_check', b'1', 1),
-        pytest.param('int_check', 12, 12, id='int_check-12_int-12_int'),
-        pytest.param('int_check', '12', 12, id='int_check-12_str-12_int'),
-        pytest.param('int_check', b'12', 12, id='int_check-12_bytes-12_int'),
-        pytest.param('float_check', 1, 1.0, id='float_check-1_int-1.0_float'),
-        pytest.param('float_check', 1.0, 1.0, id='float_check-1.0_float-1.0_float'),
-        ('float_check', Decimal(1.0), 1.0),
-        pytest.param('float_check', '1.0', 1.0, id='float_check-1.0_str-1.0_float'),
-        pytest.param('float_check', '1', 1.0, id='float_check-1_str-1.0_float'),
-        pytest.param('float_check', b'1.0', 1.0, id='float_check-1.0_bytes-1.0_float'),
-        pytest.param('float_check', b'1', 1.0, id='float_check-1_bytes-1.0_float'),
-        ('float_check', True, 1.0),
-        ('float_check', False, 0.0),
-        pytest.param('float_check', 't', ValidationError, id='float_check-t_str-ValidationError'),
-        pytest.param('float_check', b't', ValidationError, id='float_check-t_bytes-ValidationError'),
         ('uuid_check', 'ebcdab58-6eb8-46fb-a190-d07a33e9eac8', UUID('ebcdab58-6eb8-46fb-a190-d07a33e9eac8')),
         ('uuid_check', UUID('ebcdab58-6eb8-46fb-a190-d07a33e9eac8'), UUID('ebcdab58-6eb8-46fb-a190-d07a33e9eac8')),
         ('uuid_check', b'ebcdab58-6eb8-46fb-a190-d07a33e9eac8', UUID('ebcdab58-6eb8-46fb-a190-d07a33e9eac8')),
         ('uuid_check', b'\x12\x34\x56\x78' * 4, UUID('12345678-1234-5678-1234-567812345678')),
         ('uuid_check', 'ebcdab58-6eb8-46fb-a190-', ValidationError),
         ('uuid_check', 123, ValidationError),
-        ('decimal_check', 42.24, Decimal('42.24')),
-        ('decimal_check', '42.24', Decimal('42.24')),
-        ('decimal_check', b'42.24', ValidationError),
-        ('decimal_check', '  42.24  ', Decimal('42.24')),
-        ('decimal_check', Decimal('42.24'), Decimal('42.24')),
-        ('decimal_check', 'not a valid decimal', ValidationError),
-        ('decimal_check', 'NaN', ValidationError),
         ('date_check', date(2017, 5, 5), date(2017, 5, 5)),
         ('date_check', datetime(2017, 5, 5), date(2017, 5, 5)),
         pytest.param('date_check', '2017-05-05', date(2017, 5, 5), id='date_check-2017-05-05_str-2017-05-05_date'),
@@ -1733,237 +1626,6 @@ def test_sequence_strict():
     assert TypeAdapter(Sequence[int]).validate_python((), strict=True) == ()
 
 
-def test_int_validation():
-    class Model(BaseModel):
-        a: PositiveInt = None
-        b: NegativeInt = None
-        c: NonNegativeInt = None
-        d: NonPositiveInt = None
-        e: conint(gt=4, lt=10) = None
-        f: conint(ge=0, le=10) = None
-        g: conint(multiple_of=5) = None
-
-    m = Model(a=5, b=-5, c=0, d=0, e=5, f=0, g=25)
-    assert m.model_dump() == {'a': 5, 'b': -5, 'c': 0, 'd': 0, 'e': 5, 'f': 0, 'g': 25}
-
-    with pytest.raises(ValidationError) as exc_info:
-        Model(a=-5, b=5, c=-5, d=5, e=-5, f=11, g=42)
-    # insert_assert(exc_info.value.errors(include_url=False))
-    assert exc_info.value.errors(include_url=False) == [
-        {
-            'type': 'greater_than',
-            'loc': ('a',),
-            'msg': 'Input should be greater than 0',
-            'input': -5,
-            'ctx': {'gt': 0},
-        },
-        {
-            'type': 'less_than',
-            'loc': ('b',),
-            'msg': 'Input should be less than 0',
-            'input': 5,
-            'ctx': {'lt': 0},
-        },
-        {
-            'type': 'greater_than_equal',
-            'loc': ('c',),
-            'msg': 'Input should be greater than or equal to 0',
-            'input': -5,
-            'ctx': {'ge': 0},
-        },
-        {
-            'type': 'less_than_equal',
-            'loc': ('d',),
-            'msg': 'Input should be less than or equal to 0',
-            'input': 5,
-            'ctx': {'le': 0},
-        },
-        {
-            'type': 'greater_than',
-            'loc': ('e',),
-            'msg': 'Input should be greater than 4',
-            'input': -5,
-            'ctx': {'gt': 4},
-        },
-        {
-            'type': 'less_than_equal',
-            'loc': ('f',),
-            'msg': 'Input should be less than or equal to 10',
-            'input': 11,
-            'ctx': {'le': 10},
-        },
-        {
-            'type': 'multiple_of',
-            'loc': ('g',),
-            'msg': 'Input should be a multiple of 5',
-            'input': 42,
-            'ctx': {'multiple_of': 5},
-        },
-    ]
-
-
-def test_float_validation():
-    class Model(BaseModel):
-        a: PositiveFloat = None
-        b: NegativeFloat = None
-        c: NonNegativeFloat = None
-        d: NonPositiveFloat = None
-        e: confloat(gt=4, lt=12.2) = None
-        f: confloat(ge=0, le=9.9) = None
-        g: confloat(multiple_of=0.5) = None
-        h: confloat(allow_inf_nan=False) = None
-
-    m = Model(a=5.1, b=-5.2, c=0, d=0, e=5.3, f=9.9, g=2.5, h=42)
-    assert m.model_dump() == {'a': 5.1, 'b': -5.2, 'c': 0, 'd': 0, 'e': 5.3, 'f': 9.9, 'g': 2.5, 'h': 42}
-
-    assert Model(a=float('inf')).a == float('inf')
-    assert Model(b=float('-inf')).b == float('-inf')
-
-    with pytest.raises(ValidationError) as exc_info:
-        Model(a=-5.1, b=5.2, c=-5.1, d=5.1, e=-5.3, f=9.91, g=4.2, h=float('nan'))
-    # insert_assert(exc_info.value.errors(include_url=False))
-    assert exc_info.value.errors(include_url=False) == [
-        {
-            'type': 'greater_than',
-            'loc': ('a',),
-            'msg': 'Input should be greater than 0',
-            'input': -5.1,
-            'ctx': {
-                'gt': 0.0,
-            },
-        },
-        {
-            'type': 'less_than',
-            'loc': ('b',),
-            'msg': 'Input should be less than 0',
-            'input': 5.2,
-            'ctx': {
-                'lt': 0.0,
-            },
-        },
-        {
-            'type': 'greater_than_equal',
-            'loc': ('c',),
-            'msg': 'Input should be greater than or equal to 0',
-            'input': -5.1,
-            'ctx': {
-                'ge': 0.0,
-            },
-        },
-        {
-            'type': 'less_than_equal',
-            'loc': ('d',),
-            'msg': 'Input should be less than or equal to 0',
-            'input': 5.1,
-            'ctx': {
-                'le': 0.0,
-            },
-        },
-        {
-            'type': 'greater_than',
-            'loc': ('e',),
-            'msg': 'Input should be greater than 4',
-            'input': -5.3,
-            'ctx': {
-                'gt': 4.0,
-            },
-        },
-        {
-            'type': 'less_than_equal',
-            'loc': ('f',),
-            'msg': 'Input should be less than or equal to 9.9',
-            'input': 9.91,
-            'ctx': {
-                'le': 9.9,
-            },
-        },
-        {
-            'type': 'multiple_of',
-            'loc': ('g',),
-            'msg': 'Input should be a multiple of 0.5',
-            'input': 4.2,
-            'ctx': {
-                'multiple_of': 0.5,
-            },
-        },
-        {
-            'type': 'finite_number',
-            'loc': ('h',),
-            'msg': 'Input should be a finite number',
-            'input': HasRepr('nan'),
-        },
-    ]
-
-
-def test_infinite_float_validation():
-    class Model(BaseModel):
-        a: float = None
-
-    assert Model(a=float('inf')).a == float('inf')
-    assert Model(a=float('-inf')).a == float('-inf')
-    assert math.isnan(Model(a=float('nan')).a)
-
-
-@pytest.mark.parametrize(
-    ('ser_json_inf_nan', 'input', 'output', 'python_roundtrip'),
-    (
-        ('null', float('inf'), 'null', None),
-        ('null', float('-inf'), 'null', None),
-        ('null', float('nan'), 'null', None),
-        ('constants', float('inf'), 'Infinity', float('inf')),
-        ('constants', float('-inf'), '-Infinity', float('-inf')),
-        ('constants', float('nan'), 'NaN', IsFloatNan),
-    ),
-)
-def test_infinite_float_json_serialization(ser_json_inf_nan, input, output, python_roundtrip):
-    class Model(BaseModel):
-        model_config = ConfigDict(ser_json_inf_nan=ser_json_inf_nan)
-        a: float
-
-    json_string = Model(a=input).model_dump_json()
-    assert json_string == f'{{"a":{output}}}'
-    assert json.loads(json_string) == {'a': python_roundtrip}
-
-
-@pytest.mark.parametrize('value', [float('inf'), float('-inf'), float('nan')])
-def test_finite_float_validation_error(value):
-    class Model(BaseModel):
-        a: FiniteFloat
-
-    assert Model(a=42).a == 42
-    with pytest.raises(ValidationError) as exc_info:
-        Model(a=value)
-    # insert_assert(exc_info.value.errors(include_url=False))
-    assert exc_info.value.errors(include_url=False) == [
-        {
-            'type': 'finite_number',
-            'loc': ('a',),
-            'msg': 'Input should be a finite number',
-            'input': HasRepr(repr(value)),
-        }
-    ]
-
-
-def test_finite_float_config():
-    class Model(BaseModel):
-        a: float
-
-        model_config = ConfigDict(allow_inf_nan=False)
-
-    assert Model(a=42).a == 42
-    with pytest.raises(ValidationError) as exc_info:
-        Model(a=float('nan'))
-    # insert_assert(exc_info.value.errors(include_url=False))
-    assert exc_info.value.errors(include_url=False) == [
-        {
-            'type': 'finite_number',
-            'loc': ('a',),
-            'msg': 'Input should be a finite number',
-            'input': HasRepr('nan'),
-        }
-    ]
-
-
 def test_strict_bool():
     class Model(BaseModel):
         v: StrictBool
@@ -1979,52 +1641,6 @@ def test_strict_bool():
 
     with pytest.raises(ValidationError):
         Model(v=b'1')
-
-
-def test_strict_int():
-    class Model(BaseModel):
-        v: StrictInt
-
-    assert Model(v=123456).v == 123456
-
-    with pytest.raises(ValidationError, match=r'Input should be a valid integer \[type=int_type,'):
-        Model(v='123456')
-
-    with pytest.raises(ValidationError, match=r'Input should be a valid integer \[type=int_type,'):
-        Model(v=3.14159)
-
-    with pytest.raises(ValidationError, match=r'Input should be a valid integer \[type=int_type,'):
-        Model(v=True)
-
-
-@pytest.mark.parametrize(
-    ('input', 'expected_json'),
-    (
-        (9_223_372_036_854_775_807, b'9223372036854775807'),
-        (-9_223_372_036_854_775_807, b'-9223372036854775807'),
-        (1433352099889938534014333520998899385340, b'1433352099889938534014333520998899385340'),
-        (-1433352099889938534014333520998899385340, b'-1433352099889938534014333520998899385340'),
-    ),
-)
-def test_big_int_json(input, expected_json):
-    v = TypeAdapter(int)
-    dumped = v.dump_json(input)
-    assert dumped == expected_json
-    assert v.validate_json(dumped) == input
-
-
-def test_strict_float():
-    class Model(BaseModel):
-        v: StrictFloat
-
-    assert Model(v=3.14159).v == 3.14159
-    assert Model(v=123456).v == 123456
-
-    with pytest.raises(ValidationError, match=r'Input should be a valid number \[type=float_type,'):
-        Model(v='3.14159')
-
-    with pytest.raises(ValidationError, match=r'Input should be a valid number \[type=float_type,'):
-        Model(v=True)
 
 
 def test_bool_unhashable_fails():
@@ -2257,285 +1873,6 @@ def test_uuid_strict() -> None:
         assert isinstance(m.c, type(c)) and m.c == c
         assert isinstance(m.d, type(d)) and m.d == d
         assert isinstance(m.e, type(e)) and m.e == e
-
-
-pos_int_values = 'Inf', '+Inf', 'Infinity', '+Infinity'
-neg_int_values = '-Inf', '-Infinity'
-nan_values = 'NaN', '-NaN', '+NaN', 'sNaN', '-sNaN', '+sNaN'
-non_finite_values = nan_values + pos_int_values + neg_int_values
-# dirty_equals.AnyThing() doesn't work with Decimal on PyPy, hence this hack
-ANY_THING = object()
-
-
-@pytest.mark.parametrize(
-    'type_args,value,result',
-    [
-        (dict(gt=Decimal('42.24')), Decimal('43'), Decimal('43')),
-        (
-            dict(gt=Decimal('42.24')),
-            Decimal('42'),
-            [
-                {
-                    'type': 'greater_than',
-                    'loc': ('foo',),
-                    'msg': 'Input should be greater than 42.24',
-                    'input': Decimal('42'),
-                    'ctx': {'gt': Decimal('42.24')},
-                }
-            ],
-        ),
-        (dict(lt=Decimal('42.24')), Decimal('42'), Decimal('42')),
-        (
-            dict(lt=Decimal('42.24')),
-            Decimal('43'),
-            [
-                {
-                    'type': 'less_than',
-                    'loc': ('foo',),
-                    'msg': 'Input should be less than 42.24',
-                    'input': Decimal('43'),
-                    'ctx': {
-                        'lt': Decimal('42.24'),
-                    },
-                },
-            ],
-        ),
-        (dict(ge=Decimal('42.24')), Decimal('43'), Decimal('43')),
-        (dict(ge=Decimal('42.24')), Decimal('42.24'), Decimal('42.24')),
-        (
-            dict(ge=Decimal('42.24')),
-            Decimal('42'),
-            [
-                {
-                    'type': 'greater_than_equal',
-                    'loc': ('foo',),
-                    'msg': 'Input should be greater than or equal to 42.24',
-                    'input': Decimal('42'),
-                    'ctx': {
-                        'ge': Decimal('42.24'),
-                    },
-                }
-            ],
-        ),
-        (dict(le=Decimal('42.24')), Decimal('42'), Decimal('42')),
-        (dict(le=Decimal('42.24')), Decimal('42.24'), Decimal('42.24')),
-        (
-            dict(le=Decimal('42.24')),
-            Decimal('43'),
-            [
-                {
-                    'type': 'less_than_equal',
-                    'loc': ('foo',),
-                    'msg': 'Input should be less than or equal to 42.24',
-                    'input': Decimal('43'),
-                    'ctx': {
-                        'le': Decimal('42.24'),
-                    },
-                }
-            ],
-        ),
-        (dict(max_digits=2, decimal_places=2), Decimal('0.99'), Decimal('0.99')),
-        pytest.param(
-            dict(max_digits=2, decimal_places=1),
-            Decimal('0.99'),
-            [
-                {
-                    'type': 'decimal_max_places',
-                    'loc': ('foo',),
-                    'msg': 'Decimal input should have no more than 1 decimal place',
-                    'input': Decimal('0.99'),
-                    'ctx': {
-                        'decimal_places': 1,
-                    },
-                }
-            ],
-        ),
-        (
-            dict(max_digits=3, decimal_places=1),
-            Decimal('999'),
-            [
-                {
-                    'loc': ('foo',),
-                    'msg': 'Decimal input should have no more than 2 digits before the decimal point',
-                    'type': 'decimal_whole_digits',
-                    'input': Decimal('999'),
-                    'ctx': {'whole_digits': 2},
-                }
-            ],
-        ),
-        (dict(max_digits=4, decimal_places=1), Decimal('999'), Decimal('999')),
-        (dict(max_digits=20, decimal_places=2), Decimal('742403889818000000'), Decimal('742403889818000000')),
-        (dict(max_digits=20, decimal_places=2), Decimal('7.42403889818E+17'), Decimal('7.42403889818E+17')),
-        (dict(max_digits=6, decimal_places=2), Decimal('000000000001111.700000'), Decimal('000000000001111.700000')),
-        (
-            dict(max_digits=6, decimal_places=2),
-            Decimal('0000000000011111.700000'),
-            [
-                {
-                    'type': 'decimal_whole_digits',
-                    'loc': ('foo',),
-                    'msg': 'Decimal input should have no more than 4 digits before the decimal point',
-                    'input': Decimal('11111.700000'),
-                    'ctx': {'whole_digits': 4},
-                }
-            ],
-        ),
-        (
-            dict(max_digits=20, decimal_places=2),
-            Decimal('7424742403889818000000'),
-            [
-                {
-                    'type': 'decimal_max_digits',
-                    'loc': ('foo',),
-                    'msg': 'Decimal input should have no more than 20 digits in total',
-                    'input': Decimal('7424742403889818000000'),
-                    'ctx': {
-                        'max_digits': 20,
-                    },
-                },
-            ],
-        ),
-        (dict(max_digits=5, decimal_places=2), Decimal('7304E-1'), Decimal('7304E-1')),
-        (
-            dict(max_digits=5, decimal_places=2),
-            Decimal('7304E-3'),
-            [
-                {
-                    'type': 'decimal_max_places',
-                    'loc': ('foo',),
-                    'msg': 'Decimal input should have no more than 2 decimal places',
-                    'input': Decimal('7.304'),
-                    'ctx': {'decimal_places': 2},
-                }
-            ],
-        ),
-        (dict(max_digits=5, decimal_places=5), Decimal('70E-5'), Decimal('70E-5')),
-        (
-            dict(max_digits=4, decimal_places=4),
-            Decimal('70E-6'),
-            [
-                {
-                    'loc': ('foo',),
-                    'msg': 'Decimal input should have no more than 4 digits in total',
-                    'type': 'decimal_max_digits',
-                    'input': Decimal('0.00007'),
-                    'ctx': {'max_digits': 4},
-                }
-            ],
-        ),
-        *[
-            (
-                dict(decimal_places=2, max_digits=10, allow_inf_nan=False),
-                value,
-                [
-                    {
-                        'loc': ('foo',),
-                        'msg': 'Input should be a finite number',
-                        'type': 'finite_number',
-                        'input': value,
-                    }
-                ],
-            )
-            for value in non_finite_values
-        ],
-        *[
-            (
-                dict(decimal_places=2, max_digits=10, allow_inf_nan=False),
-                Decimal(value),
-                [
-                    {
-                        'loc': ('foo',),
-                        'msg': 'Input should be a finite number',
-                        'type': 'finite_number',
-                        'input': ANY_THING,
-                    }
-                ],
-            )
-            for value in non_finite_values
-        ],
-        (
-            dict(multiple_of=Decimal('5')),
-            Decimal('42'),
-            [
-                {
-                    'type': 'multiple_of',
-                    'loc': ('foo',),
-                    'msg': 'Input should be a multiple of 5',
-                    'input': Decimal('42'),
-                    'ctx': {'multiple_of': Decimal('5')},
-                }
-            ],
-        ),
-    ],
-)
-@pytest.mark.parametrize('mode', ['Field', 'condecimal', 'optional'])
-def test_decimal_validation(mode, type_args, value, result):
-    if mode == 'Field':
-
-        class Model(BaseModel):
-            foo: Decimal = Field(**type_args)
-
-    elif mode == 'optional':
-
-        class Model(BaseModel):
-            foo: Decimal | None = Field(**type_args)
-
-    else:
-
-        class Model(BaseModel):
-            foo: condecimal(**type_args)
-
-    if not isinstance(result, Decimal):
-        with pytest.raises(ValidationError) as exc_info:
-            m = Model(foo=value)
-            print(f'unexpected result: {m!r}')
-        # debug(exc_info.value.errors(include_url=False))
-        # dirty_equals.AnyThing() doesn't work with Decimal on PyPy, hence this hack
-        errors = exc_info.value.errors(include_url=False)
-        if result[0].get('input') is ANY_THING:
-            for e in errors:
-                e['input'] = ANY_THING
-        assert errors == result
-        # assert exc_info.value.json().startswith('[')
-    else:
-        assert Model(foo=value).foo == result
-
-
-@pytest.fixture(scope='module', name='AllowInfModel')
-def fix_allow_inf_model():
-    class Model(BaseModel):
-        v: condecimal(allow_inf_nan=True)
-
-    return Model
-
-
-@pytest.mark.parametrize(
-    'value,result',
-    [
-        (Decimal('42'), 'unchanged'),
-        *[(v, 'is_nan') for v in nan_values],
-        *[(v, 'is_pos_inf') for v in pos_int_values],
-        *[(v, 'is_neg_inf') for v in neg_int_values],
-    ],
-)
-def test_decimal_not_finite(value, result, AllowInfModel):
-    m = AllowInfModel(v=value)
-    if result == 'unchanged':
-        assert m.v == value
-    elif result == 'is_nan':
-        assert m.v.is_nan(), m.v
-    elif result == 'is_pos_inf':
-        assert m.v.is_infinite() and m.v > 0, m.v
-    else:
-        assert result == 'is_neg_inf'
-        assert m.v.is_infinite() and m.v < 0, m.v
-
-
-def test_decimal_invalid():
-    with pytest.raises(SchemaError, match='allow_inf_nan=True cannot be used with max_digits or decimal_places'):
-
-        class Model(BaseModel):
-            v: condecimal(allow_inf_nan=True, max_digits=4)
 
 
 @pytest.mark.parametrize('value,result', (('/test/path', Path('/test/path')), (Path('/test/path'), Path('/test/path'))))
@@ -2836,140 +2173,6 @@ def test_ser_path_incorrect() -> None:
     ta = TypeAdapter(Path)
     with pytest.warns(UserWarning, match='serialized value may not be as expected.'):
         ta.dump_python(123)
-
-
-def test_number_gt():
-    class Model(BaseModel):
-        a: conint(gt=-1) = 0
-
-    assert Model(a=0).model_dump() == {'a': 0}
-
-    with pytest.raises(ValidationError) as exc_info:
-        Model(a=-1)
-    # insert_assert(exc_info.value.errors(include_url=False))
-    assert exc_info.value.errors(include_url=False) == [
-        {
-            'type': 'greater_than',
-            'loc': ('a',),
-            'msg': 'Input should be greater than -1',
-            'input': -1,
-            'ctx': {'gt': -1},
-        }
-    ]
-
-
-def test_number_ge():
-    class Model(BaseModel):
-        a: conint(ge=0) = 0
-
-    assert Model(a=0).model_dump() == {'a': 0}
-
-    with pytest.raises(ValidationError) as exc_info:
-        Model(a=-1)
-    # insert_assert(exc_info.value.errors(include_url=False))
-    assert exc_info.value.errors(include_url=False) == [
-        {
-            'type': 'greater_than_equal',
-            'loc': ('a',),
-            'msg': 'Input should be greater than or equal to 0',
-            'input': -1,
-            'ctx': {'ge': 0},
-        }
-    ]
-
-
-def test_number_lt():
-    class Model(BaseModel):
-        a: conint(lt=5) = 0
-
-    assert Model(a=4).model_dump() == {'a': 4}
-
-    with pytest.raises(ValidationError) as exc_info:
-        Model(a=5)
-    # insert_assert(exc_info.value.errors(include_url=False))
-    assert exc_info.value.errors(include_url=False) == [
-        {
-            'type': 'less_than',
-            'loc': ('a',),
-            'msg': 'Input should be less than 5',
-            'input': 5,
-            'ctx': {'lt': 5},
-        }
-    ]
-
-
-def test_number_le():
-    class Model(BaseModel):
-        a: conint(le=5) = 0
-
-    assert Model(a=5).model_dump() == {'a': 5}
-
-    with pytest.raises(ValidationError) as exc_info:
-        Model(a=6)
-    # insert_assert(exc_info.value.errors(include_url=False))
-    assert exc_info.value.errors(include_url=False) == [
-        {
-            'type': 'less_than_equal',
-            'loc': ('a',),
-            'msg': 'Input should be less than or equal to 5',
-            'input': 6,
-            'ctx': {'le': 5},
-        }
-    ]
-
-
-@pytest.mark.parametrize('value', (10, 100, 20))
-def test_number_multiple_of_int_valid(value):
-    class Model(BaseModel):
-        a: conint(multiple_of=5)
-
-    assert Model(a=value).model_dump() == {'a': value}
-
-
-@pytest.mark.parametrize('value', [1337, 23, 6, 14])
-def test_number_multiple_of_int_invalid(value):
-    class Model(BaseModel):
-        a: conint(multiple_of=5)
-
-    with pytest.raises(ValidationError) as exc_info:
-        Model(a=value)
-    # insert_assert(exc_info.value.errors(include_url=False))
-    assert exc_info.value.errors(include_url=False) == [
-        {
-            'type': 'multiple_of',
-            'loc': ('a',),
-            'msg': 'Input should be a multiple of 5',
-            'input': value,
-            'ctx': {'multiple_of': 5},
-        }
-    ]
-
-
-@pytest.mark.parametrize('value', [0.2, 0.3, 0.4, 0.5, 1])
-def test_number_multiple_of_float_valid(value):
-    class Model(BaseModel):
-        a: confloat(multiple_of=0.1)
-
-    assert Model(a=value).model_dump() == {'a': value}
-
-
-@pytest.mark.parametrize('value', [0.07, 1.27, 1.003])
-def test_number_multiple_of_float_invalid(value):
-    class Model(BaseModel):
-        a: confloat(multiple_of=0.1)
-
-    with pytest.raises(ValidationError) as exc_info:
-        Model(a=value)
-    # insert_assert(exc_info.value.errors(include_url=False))
-    assert exc_info.value.errors(include_url=False) == [
-        {
-            'type': 'multiple_of',
-            'loc': ('a',),
-            'msg': 'Input should be a multiple of 0.1',
-            'input': value,
-            'ctx': {'multiple_of': 0.1},
-        }
-    ]
 
 
 def test_new_type_success():
@@ -5621,17 +4824,6 @@ def test_get_pydantic_core_schema_marker_unrelated_type() -> None:
     ta = TypeAdapter(Annotated[int, Marker(2), Marker(3)])
 
     assert ta.validate_python('1') == 3
-
-
-def test_decimal_float_precision() -> None:
-    """https://github.com/pydantic/pydantic/issues/6807"""
-    ta = TypeAdapter(Decimal)
-    assert ta.validate_json('1.1') == Decimal('1.1')
-    assert ta.validate_python(1.1) == Decimal('1.1')
-    assert ta.validate_json('"1.1"') == Decimal('1.1')
-    assert ta.validate_python('1.1') == Decimal('1.1')
-    assert ta.validate_json('1') == Decimal('1')
-    assert ta.validate_python(1) == Decimal('1')
 
 
 def test_union_tags_in_errors():
